pyproject.toml
[tool.poetry]
name = "pl_worker"
version = "0.1.0"
description = ""
authors = ["Jesse Schoepfer <jelloeater@gmail.com>"]
readme = "README.md"
#homepage = "https://github.com/Jelloeater/xx"
#keywords = ["hubitat", "makerapi","requests"]
#classifiers = ["Development Status :: 5 - Production/Stable",
# "Environment :: No Input/Output (Daemon)",
# "Topic :: Home Automation",
# "Intended Audience :: Developers",
# "Operating System :: OS Independent",
#]

[tool.poetry.dependencies]
python = "^3.10"

hubitatcontrol = {git = "https://github.com/Jelloeater/hubitatcontrol.git"}
extcolors = "^1.0.0"
google-images-search = "^1.4.6"
fastapi ="*"
uvicorn ={version="*", extras = ["standard"]}

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = "test/"

[tool.black]
line-length = 120

[tool.isort]
multi_line_output = 3
include_trailing_comma = "True"
force_grid_wrap = 0
use_parentheses = "True"
line_length = 120


[tool.poetry.group.dev.dependencies]
setuptools = "*"
pytest = "*"
pytest-cov = "*"
python-dotenv = "*"
platformdirs = "*"
pre-commit = "*"
pytest-pycharm = "*"
vulture = "*"
isort = "*"
pdoc3 = "*"
radon = "*"
xenon = "*"
bandit = "*"
black = "*"
pytest-xprocess ="*"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
